Rearranged reg-validation.php, moved logic code to register php file, changed register form action, added option for gender selection
